organize tools and add README for FairRoulette
