+++ tools/fairroulette/cmd/fr-admin/main.go
