+++ tools/wallet/cmd/wallet/main.go
